Split DevicesController: extract create_device
Move create_device into a sibling mutations_create.py as a
free function taking controller as the first arg. The
controller keeps a thin bound-method delegate so WS dispatch
and tests calling controller.create_device(...) still
resolve unchanged.

Drops parse_platform_from_yaml + remove_device_metadata
imports from controller.py (only used by the moved code);
StorageJSON / resolve_storage_path / set_device_metadata /
friendly_name_slugify stay since other still-on-controller
methods (rename_device, clone_device, _persist_storage_version)
also use them.

stub_create_device_metadata_helpers fixture's stub patches
re-pointed at mutations_create so they keep intercepting the
create_device flow. The test_branches_coverage.py:674 patch
on controller.resolve_storage_path stays since it tests
_persist_storage_version_async which still lives on the
controller.

Pure refactor; behaviour preserved. Drops ~175 lines from
controllers/devices/controller.py (now 1478).
